requirements.txt
PyYAML
requests
reportlab
python-dotenv
openai
pychain
